% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{HDdata}
\alias{HDdata}
\title{Statlog Heart Data Set}
\source{
UCI Machine Learning Repository, Statlog (Heart) Data Set.
}
\description{
This dataset from UCI Machine Learning repository is a 
heart disease database similar to the database already present in the 
repository (Heart Disease databases) but in a slightly different form. 
The 14 attributes are:
\itemize{
\item age
\item sex (0=female, 1=male),
\item chest pain type ( 1=typical angina, 2=atypical angina, 3=non-anginal pain, 4:=asymptomatic),
\item resting blood pressure in  mm Hg
\item serum cholestoral in mg/dl
\item fasting blood sugar > 120 mg/dl (1=True, 0=False)
\item resting electrocardiographic results (0=normal,1=ST-T wave abnormality 
, 2= elevation or depression of > 0.05 mV)
\item maximum heart rate achieved
\item exercise induced angina (1=Yes, 0=No)
\item oldpeak = ST depression induced by exercise relative to rest
\item the slope of the peak exercise ST segment (1=upsloping, 2=flat,3=downsloping)
\item number of major vessels (0-3) colored by flourosopy
\item thal: 3 = normal; 6 = fixed defect; 7 = reversable defect
\item heart disease : 1=absence, 2=presence
}
}
