% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optVardim.R
\name{optVardim}
\alias{optVardim}
\title{Optimal projected variance of non orthogonal principal components}
\usage{
optVardim(B, Z)
}
\arguments{
\item{B}{a numerical data matrix  (usually centered and/or scaled, see details).}

\item{Z}{a numerical loading matrix.}
}
\value{
Returns the optimal projected variance of each principal components.
}
\description{
This function calculates the variance explained by
not necessarily orthogonal principal components, using the optimal projected 
variance (optVar).
}
\details{
The principal components are defined by
Y=BZ where B is the centered and/or scaled data matrix and Z is the sparse 
loading matrix. The argument \code{B} must then be consistent with the 
pre-processing step in \code{\link{sparsePCA}} and \code{\link{groupsparsePCA}}.
The loading vectors in \code{Z} must be unique norm and linearly independant.
}
\examples{
data(protein)
B <- sparsePCA(protein, 2, c(0.5,0.5))$B
Z <- sparsePCA(protein, 2, c(0.5,0.5))$Z
optVardim(B,Z)
 
}
\references{
M. Chavent and G. Chavent, Optimal projected variance group-sparse
block PCA, submitted, 2020.
}
\seealso{
\code{\link{pev}}, \code{\link{explainedVar}}, 
\code{\link{sparsePCA}}, \code{\link{groupsparsePCA}}
}
