% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pev.R
\name{pev}
\alias{pev}
\title{Proportion of variance explained by non orthogonal principal components}
\usage{
pev(x)
}
\arguments{
\item{x}{a object of class \code{sparsePCA}.}
}
\value{
The proportion of variance explained by each principal components.
}
\description{
This function calculates the proportion of variance explained by
not necessarily orthogonal principal components, using the optimal projected
variance (optVar).
}
\details{
The object \code{x} is the output of \code{\link{sparsePCA}} or 
\code{\link{groupsparsePCA}}. The  variance of the not necessarily 
orthogonal principal components in \code{x}
is calculated with the optimal projected variance definition (see Chavent and Chavent, 2020)
and is divided by the total variance of the data to get a proportion of
explained variance (pev).
}
\examples{
data(protein)
x <- sparsePCA(protein, 2, c(0.5,0.5))
pev(x)
x <- groupsparsePCA(protein, 2, index=1:ncol(protein), c(0.5,0.5))
pev(x)
}
\references{
M. Chavent and G. Chavent, Optimal projected variance group-sparse
block PCA, submitted, 2020.
}
\seealso{
\code{\link{optVardim}}, \code{\link{explainedVar}}, 
\code{\link{sparsePCA}}, \code{\link{groupsparsePCA}},
}
