% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critEval.R
\name{RVcoef}
\alias{RVcoef}
\title{RV coefficient}
\usage{
RVcoef(X, Y)
}
\arguments{
\item{X}{a matrix of dimension n times p}

\item{Y}{a matrix of dimension n times q}
}
\description{
The RV coefficient is a measures the proximity between two matrices (subspaces spanned by the columns ofthe matrices). This measure
is normalized and takes its values between 0 and 1.
}

