% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simusparsePCA.R
\name{simusparsePCA}
\alias{simusparsePCA}
\title{Simulation from a sparse PCA model}
\usage{
simusparsePCA(n = 30, Ztrue, valp, seed = FALSE)
}
\arguments{
\item{n}{the number of observations.}

\item{Ztrue}{a numerical matrix of dimension p times m. The columns of Ztrue are m sparse and othogonal vectors. 
They will be normalized and used as the m leading eigenvectors of the covariance matrix.}

\item{valp}{a numerical vector of size p. The values are the eigenvalues of the covariance matrix.}

\item{seed}{a seed value to replicate the simulation. By default, seed=FALSE.}
}
\value{
Returns a data matrix A of dimension n times p.
}
\description{
Simulation of a data matrix  drawn from a sparse PCA model. 
The data matrix A of dimension n times p is drawn from a centered multivariate normal 
distribution. The m < p leading eigenvectors of the covariance matrix of this normal distribution are sparse.
}
\details{
This function generates data with a covariance matrix C having sparse eigenvectors V. C is synthesized
through the eigenvalue decomposition C=VDV^T where D=diag(valp). The first m columns of V are the sparse and orthogonal 
columns of Ztrue.  The remaining p-m eigenvectors are not specified to be sparse and are generated as follows. A matrix U of 
dimension n times (p-m) is randomly dawn from U(0,1) and V=Q from the QR decomposition  [Ztrue,U]=QR. The data matrix A is 
then generated by drawing n samples from a zero-mean distribution with covariance matrix C.
}
\examples{
# Example from Journee & al. 2010
 v1 <- c(rep(1/sqrt(10),10),rep(0,490))
 v2 <- c(rep(0,10), rep(1/sqrt(10),10),rep(0,480))
 valp <- c(400,300,rep(1,498))
 n <- 100
 A <- simusparsePCA(n,cbind(v1,v2),valp,seed=1)
 svd(A/sqrt(n))$d[1:3]^2 #eigenvalues of the empirical covariance matrix.
 
# Example from Shen & Huang 2008
 v1 <- c(1,1,1,1,0,0,0,0,0.9,0.9)
 v2 <- c(0,0,0,0,1,1,1,1,-0.3,0.3)
 valp <- c(200,100,50,50,6,5,4,3,2,1)
 n <- 100
 A <- simusparsePCA(n,cbind(v1,v2),valp,seed=1)
 svd(A/sqrt(n))$d^2 #eigenvalues of the empirical covariance matrix.
 
# Example from Chavent & Chavent 2017
 valp <- c(c(200,180,150,130),rep(1,16))
 data("Ztrue")
 n <- 100
 A <- simusparsePCA(n,Ztrue,valp,seed=1)
 svd(A/sqrt(n))$d^2 #eigenvalues of the empirical covariance matrix.
}
\references{
\itemize{
\item M. Chavent and G. Chavent, 
\item M. Journee and al., Generalized Power Method for Sparse Principal Component Analysis, Journal of Machine Learning Research 11 (2010) 517-553.
\item H. Shen and J.Z. Huang, Journal of Multivariate Analysis 99 (2008) 1015-1034.
}
}
