% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critEval.R
\name{Vol}
\alias{Vol}
\title{Volume coefficient}
\usage{
Vol(B, Z)
}
\arguments{
\item{B}{a n times p (centered or standardized) data matrix.}

\item{Z}{a p times m matrix of sparse loadings.}
}
\description{
The volume of the parallepipede constructed on the columns of the matrix Y=BZ. This volume measures the orthogonality of the columns of Y.
It is normalized and takes its values between 0 and 1.
}
