% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsePCA.R
\name{Ztrue}
\alias{Ztrue}
\title{Ztrue data}
\format{A data  matrix with 20 rows and 4 columns.}
\source{
Chavent & Chavent (2017).
}
\description{
The four columns are specified leading eigenvectors of a covariance matrix used to
simulate data from a group-sparse PCA model : the 20 variables are organized in
5 groups of 4 variables.
}

