% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optVardim.R
\name{optVardim}
\alias{optVardim}
\title{Explained variance of sparse principal components}
\usage{
optVardim(B, Z)
}
\arguments{
\item{B}{a n times p  (centered or standardized) data matrix.}

\item{Z}{a p times m matrix of sparse loadings.}
}
\value{
Returns the explained variance of each principal components.
}
\description{
In PCA and sparse PCA, the principal components are
the columns of the matrix Y=BZ where B is a n times p centered (or standardized) data matrix, and Z
is the p times m matrix of loadings. In sparse PCA, the loadings are not necessarly orthogonal and 
the principal components can be correlated. The definition of the variance explained by each principal
components must then be modified. This function implements the 'optimal variance' (optVar) definition of 'explained variance'.
}
\details{
The m loadings vectors in Z must be unique norm and linearly independant.
}
\examples{
# Example from Shen & Huang 2008
 v1 <- c(1,1,1,1,0,0,0,0,0.9,0.9)
 v2 <- c(0,0,0,0,1,1,1,1,-0.3,0.3)
 valp <- c(200,100,50,50,6,5,4,3,2,1)
 A <- simusparsePCA(50,cbind(v1,v2),valp,seed=1)
 Z <- sparsePCA(A,2,c(0.5,0.5))$Z #deflation algo
 B <- sparsePCA(A,2,c(0.5,0.5))$B
 optVardim(B,Z)
}
