% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pev.R
\name{pev}
\alias{pev}
\title{Proportion of explained variance of sparse principal components}
\usage{
pev(x)
}
\arguments{
\item{x}{a object of class \code{sparsePCA}.}
}
\value{
\item{pev}{the proportion of explained variance of the m principal components i.e. the optimal variance (optVar) divided 
the total variance of the data (the Frobenius norm of A).}
}
\description{
The proportion of explained variance of the m principal components i.e. the optimal variance (optVar) divided by
the total variance of the data (the Frobenius norm of A).
}
\examples{
data(protein)
x <- sparsePCA(protein,2,c(0.5,0.5))
pev(x)
x <- groupsparsePCA(protein,2,index=1:ncol(protein),c(0.5,0.5))
pev(x)
}
\references{
\itemize{
\item M. Chavent & G. Chavent,
}
}
