% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsePCA.R
\name{sparsePCA}
\alias{sparsePCA}
\title{Sparse PCA}
\usage{
sparsePCA(A, m, lambda, block = 1, mu = 1/1:m, post = FALSE,
  center = TRUE, scale = TRUE, iter_max = 1000, epsilon = 1e-04)
}
\arguments{
\item{A}{a n times p  numerical data matrix}

\item{m}{number of components}

\item{lambda}{a vector of dimension m with reduced sparsity parameters (in relative value with respect to the theoretical upper bound). 
Each reduced sparsity parameter is a value between 0 and 1.}

\item{block}{either 0 or 1. block==0 means that deflation is used if more than one component need to be computed.
A block algorithm is otherwise used, that computes m components at once. By default, block=1.}

\item{mu}{vector of dimension m with the mu parameters (required for the block algorithms only).
By default, mu_j=1/j.}

\item{post}{If TRUE the loadings are obtained with post-processing.}

\item{center}{a logical value indicating whether the variables should be shifted to be zero centered.}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance.}

\item{iter_max}{maximum number of admissible iterations.}

\item{epsilon}{accuracy of the stopping criterion.}
}
\value{
\item{Z}{the p times m matrix that contains the m sparse loading vectors}

\item{Y}{the n times m matrix that contains the m principal components}

\item{B}{the data matrix centered (if center=TRUE) and scaled (if scale=TRUE)}
}
\description{
Sparse principal component analysis (PCA). Three algorithms are proposed for the determination of the
sparse loadings : on based on a deflation procedure and two based on block optimisation.
}
\details{
The principal components are given by Y=BZ  where B is the matrix A which has been centered (if scale=FALSE) or standardized (if scale=TRUE) and Z is the matrix of the sparse loading vectors.
}
\examples{
# Example from Shen & Huang 2008
 v1 <- c(1,1,1,1,0,0,0,0,0.9,0.9)
 v2 <- c(0,0,0,0,1,1,1,1,-0.3,0.3)
 valp <- c(200,100,50,50,6,5,4,3,2,1)
 n <- 50
 A <- simuPCA(n,cbind(v1,v2),valp,seed=1)
 # Three algorithms
 Z <- sparsePCA(A,2,c(0.5,0.5),block=0)$Z #deflation
 Z <- sparsePCA(A,2,c(0.5,0.5),block=1)$Z #block different mu
 Z <- sparsePCA(A,2,c(0.5,0.5),block=1,mu=c(1,1))$Z #block same mu
 
 # Example of the protein data
 data("protein")
 Z <- sparsePCA(protein,2,c(0.5,0.5))$Z #block different mu
 
}
\references{
\itemize{
\item M. Chavent and G. Chavent, Group-sparse block PCA and explained variance, arXiv:1705.00461
\item M. Journee and al., Generalized Power Method for Sparse Principal Component Analysis, Journal of Machine Learning Research 11 (2010) 517-553.
\item H. Shen and J.Z. Huang, Journal of Multivariate Analysis 99 (2008) 1015-1034.
}
}
\seealso{
\code{\link{groupsparsePCA}}, \code{\link{pev}}
}

