% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Ztrue}
\alias{Ztrue}
\title{Ztrue data}
\format{
A numerical matrix of size 20 by 4.
}
\source{
M. Chavent and G. Chavent, Optimal projected variance group-sparse 
block PCA, submitted, 2020.
}
\description{
This dataset is a group-sparse loading matrix. The columns 
of this matrix are the four first eigenvectors of a covariance matrix that 
can be used in \code{\link{simuPCA}} to simulate data from a PCA model.
}
